fix(codex): address websocket-cached review findings
- Close one-shot sockets on send failure; invalidate cache entries only for cached leases
- Skip connection cache for Generated correlation IDs; align body after 401 retry
- Fingerprint stable auth context without bearer token; busy acquire wins over mismatch/expiry
- Treat response.done as terminal only with valid response id; avoid double-close after invalidate
- Inject Clock for max-age expiry; simplify worker shutdown to closeAll
- Sync project settings TTL defaults; restore privacy-safe request summary logging

+++ src/Platform/Bridge/OpenAICodex/CodexWebSocketCacheEntry.php
@@ -18,7 +18,7 @@ final class CodexWebSocketCacheEntry
     public function __construct(
         public readonly WebsocketConnection $connection,
         public readonly CodexWebSocketCompatibilityFingerprint $identity,
-        public readonly float $createdAt,
+        public readonly int $createdAt,
         public readonly CodexWebSocketCacheSettings $settings,
     ) {
     }
